Fixed up initilization re:bootstrap and jquery

+++ init.sh
@@ -17,31 +17,50 @@ then
 	read -rsp $'Press any key to continue, or CTRL+C to cancel.\n' -n1 key
 
 	# Download and install lib::Local to the ./perl5 directory
+	$TMPDIR=$(mktemp -d)
+	$INSTDIR=$(pwd)
+	cd $TMPDIR
 	wget http://search.cpan.org/CPAN/authors/id/H/HA/HAARG/local-lib-2.000018.tar.gz
 	tar -xzf local-lib-2.000018.tar.gz
 	cd local-lib-2.000018
-	perl Makefile.PL --bootstrap=../perl5
+	perl Makefile.PL --bootstrap=$INSTDIR/perl5
 	make test && make install
 
 	# Cleanup the lib::local install
 	cd ..
 	rm -rf local-lib-2.000018
 	rm local-lib-2.000018.tar.gz
+	cd $INSTDIR
+	rmdir $TEMPDIR
 fi
 
 # Setup lib::local for use
 eval "$(perl -Iperl5/lib/perl5 -Mlocal::lib=perl5)"
 
 # intall the modules from cpan
-ctpan Dancer2 Template File::Slurper DBD::SQLite
+cpan Dancer2 Template File::Slurper DBD::SQLite
 
 # Install bootstrap
-wget https://github.com/twbs/bootstrap/releases/download/v3.3.6/bootstrap-3.3.6-dist.zip
-unzip bootstrap-3.3.6-dist.zip
-cp -R bootstrap-3.3.6-dist/fonts public/
-cp -R bootstrap-3.3.6-dist/css/bootstrap.min.css bootstrap-3.3.6-dist/css/bootstrap.min.css.map public/css
-cp -R bootstrap-3.3.6-dist/js/bootstrap.min.js public/javascripts
-rm -rf bootstrap-3.3.6-dist bootstrap-3.3.6-dist.zip
+if [ ! -e public/css/bootstrap.min.css ]
+then
+	echo Installing bootstrap
+	$TMPDIR=$(mktemp -d)
+	$INSTDIR=$(pwd)
+	cd $TMPDIR
+
+	wget https://github.com/twbs/bootstrap/releases/download/v3.3.6/bootstrap-3.3.6-dist.zip
+	unzip bootstrap-3.3.6-dist.zip
+	cp -R bootstrap-3.3.6-dist/fonts $INSTDIR/public/
+	cp bootstrap-3.3.6-dist/css/bootstrap.min.css bootstrap-3.3.6-dist/css/bootstrap.min.css.map $INSTDIR/public/css
+	cp -R bootstrap-3.3.6-dist/js/bootstrap.min.js $INSTDIR/public/javascripts
+	rm -rf bootstrap-3.3.6-dist bootstrap-3.3.6-dist.zip
+	cd $INSTDIR
+	rm $TEMPDIR
+fi
 
 # Install Jquery
-wget -P public/javascripts https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js
+if [ ! -e public/javascripts/jquery.min.js ]
+then
+	echo Installing jquery
+	wget https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js -O public/javascripts/jquery.min.js
+fi
